Test Case Files

+++ User-Service/User-Service/src/main/java/com/user/dto/UserDto.java
@@ -15,7 +15,7 @@ public class UserDto {
 
     private String userName;
 
-    private int mobileNumber;
+    private long mobileNumber;
 
     private String emailId;
